+++ src/commands/expressions/update/add.ts
@@ -2,7 +2,7 @@ import type { NativeNumber, NativeSet } from "../../../types.js";
 import type { AttributeNames } from "../../attributes/names.js";
 import type { AttributeValues } from "../../attributes/values.js";
 import { Path, type RawPath } from "../operands/path.js";
-import type { Value } from "../operands/value.js";
+import { type RawValue, Value } from "../operands/value.js";
 import type { IUpdateAction, UpdateAction } from "./action.js";
 import type { UpdateExpressionClauses } from "./clauses.js";
 
@@ -48,14 +48,18 @@ export class AddAction implements IUpdateAction {
  * This action only supports numbers and sets as values.
  *
  * @param rawPath - The attribute path to modify.
- * @param value - The value to add to the attribute.
+ * @param rawVAlue - The value to add to the attribute.
  *
  * @returns An {@link AddAction} that will add the value to the attribute at the specified path.
  *
  * @see https://docs.aws.amazon.com/amazondynamodb/latest/developerguide/Expressions.UpdateExpressions.html#Expressions.UpdateExpressions.ADD
  */
 
-export function add(rawPath: RawPath, value: Value<NumberOrSet>): UpdateAction {
+export function add(
+  rawPath: RawPath,
+  rawVAlue: RawValue<NumberOrSet>,
+): UpdateAction {
   const path = Path.normalize(rawPath);
+  const value = Value.normalize(rawVAlue);
   return AddAction.from({ path, value });
 }