+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.30.0] - 2025-08-05
+
+### Added
+
+- Support for `RawPath` and value normalization. Very similar to the change in version [0.29.0],
+but applied to value operands. The APIs that now support raw values are the `add` and `delete`
+update actions.
+
+
 ## [0.29.0] - 2025-08-04
 
 ### Added